Remove SecretBindings support
