+++ .gitignore
@@ -1,22 +1,2 @@
-.DS_Store
-
-# Dependency directories
-testing/e2e/skr/node_modules/
-
-testing/e2e/skr/junit_keb-e2e-skr.xml
-testing/e2e/skr/config.yaml
-
-.dev-local
-.idea/
-
-cmd/broker/catalog-test.json
-
-bin
-
-vendor
-debug
-
-# Python
-__pycache__/
-.pytest_cache/
-docs/superpowers/
+# Created by venv; see https://docs.python.org/3/library/venv.html
+*
