+++ cmd/broker/deprovisioning.go
@@ -3,10 +3,7 @@ package main
 import (
 	"context"
 	"log/slog"
-	"strings"
 	"time"
-
-	"github.com/kyma-project/kyma-environment-broker/internal/process/steps"
 
 	"k8s.io/client-go/dynamic"
 
@@ -21,8 +18,6 @@ import (
 func NewDeprovisioningProcessingQueue(ctx context.Context, workersAmount int, deprovisionManager *process.StagedManager,
 	cfg *Config, db storage.BrokerStorage,
 	k8sClientProvider K8sClientProvider, kcpClient client.Client, configProvider config.Provider, gardenerClient dynamic.Interface, gardenerNamespace string, logs *slog.Logger) *process.Queue {
-
-	useCredentialsBinding := strings.ToLower(cfg.SubscriptionGardenerResource) == credentialsBinding
 
 	deprovisioningSteps := []struct {
 		disabled bool
@@ -47,14 +42,7 @@ func NewDeprovisioningProcessingQueue(ctx context.Context, workersAmount int, de
 			step: deprovisioning.NewCheckRuntimeResourceDeletionStep(db, kcpClient, cfg.StepTimeouts.CheckRuntimeResourceDeletion),
 		},
 		{
-			disabled: useCredentialsBinding,
-			step: steps.NewHolderStep(cfg.HoldHapSteps,
-				deprovisioning.NewFreeSubscriptionStep(db.Operations(), db.Instances(), gardenerClient, gardenerNamespace)),
-		},
-		{
-			disabled: !useCredentialsBinding,
-			step: steps.NewHolderStep(cfg.HoldHapSteps,
-				deprovisioning.NewFreeCredentialsBindingStep(db.Operations(), db.Instances(), gardenerClient, gardenerNamespace)),
+			step: deprovisioning.NewFreeCredentialsBindingStep(db.Operations(), db.Instances(), gardenerClient, gardenerNamespace),
 		},
 		{
 			step: deprovisioning.NewArchivingStep(db),
