+++ cmd/broker/provisioning_test.go
@@ -437,7 +437,6 @@ func TestProvisioning_HappyPathAWS(t *testing.T) {
 func TestProvisioning_CredentialsBindings(t *testing.T) {
 	// given
 	cfg := fixConfig()
-	cfg.SubscriptionGardenerResource = "CredentialsBinding"
 	cfg.ProvidersConfigurationFilePath = providersZonesDiscovery
 	suite := NewBrokerSuiteTest(t, WithConfig(cfg))
 	defer suite.TearDown()
@@ -3605,7 +3604,6 @@ func TestProvisioning_MultiHyperscalerAccounts(t *testing.T) {
 	t.Run("disabled feature uses single-account behavior", func(t *testing.T) {
 		// given
 		cfg := fixConfig()
-		cfg.SubscriptionGardenerResource = credentialsBinding
 		cfg.HapMultiHyperscalerAccount.AllowedGlobalAccounts = []string{}
 		cfg.HapMultiHyperscalerAccount.Limits.Default = 100
 		cfg.HapMultiHyperscalerAccount.Limits.AWS = 2
@@ -3631,7 +3629,6 @@ func TestProvisioning_MultiHyperscalerAccounts(t *testing.T) {
 	t.Run("GA not in allowed global accounts uses single-account behavior", func(t *testing.T) {
 		// given
 		cfg := fixConfig()
-		cfg.SubscriptionGardenerResource = credentialsBinding
 		cfg.HapMultiHyperscalerAccount.AllowedGlobalAccounts = []string{"whitelisted-ga-001"}
 		cfg.HapMultiHyperscalerAccount.Limits.Default = 100
 		cfg.HapMultiHyperscalerAccount.Limits.AWS = 2
@@ -3657,7 +3654,6 @@ func TestProvisioning_MultiHyperscalerAccounts(t *testing.T) {
 	t.Run("enabled for specific GA selects most populated account below limit", func(t *testing.T) {
 		// given
 		cfg := fixConfig()
-		cfg.SubscriptionGardenerResource = credentialsBinding
 		cfg.HapMultiHyperscalerAccount.AllowedGlobalAccounts = []string{"multi-account-ga-001"}
 		// Set low limit to force rotation to multiple bindings
 		cfg.HapMultiHyperscalerAccount.Limits.Default = 100
@@ -3692,7 +3688,6 @@ func TestProvisioning_MultiHyperscalerAccounts(t *testing.T) {
 	t.Run("wildcard allowed global accounts enables feature for all GAs", func(t *testing.T) {
 		// given
 		cfg := fixConfig()
-		cfg.SubscriptionGardenerResource = credentialsBinding
 		cfg.HapMultiHyperscalerAccount.AllowedGlobalAccounts = []string{"*"}
 		cfg.HapMultiHyperscalerAccount.Limits.Default = 100
 		cfg.HapMultiHyperscalerAccount.Limits.AWS = 2
