+++ cmd/broker/suite_test.go
@@ -227,7 +227,6 @@ func fixConfig() *Config {
 		PlansConfigurationFilePath:              "testdata/plans.yaml",
 		RuntimeConfigurationConfigMapName:       "keb-runtime-config",
 		QuotaWhitelistedSubaccountsFilePath:     "testdata/quota_whitelist.yaml",
-		SubscriptionGardenerResource:            "secretbinding",
 		MachinesAvailabilityEndpoint:            true,
 		GvisorWhitelistedGlobalAccountsFilePath: "testdata/gvisor_whitelist.yaml",
 	}
