+++ common/gardener/client.go
@@ -47,16 +47,6 @@ func (c *Client) GetSecret(namespace, name string) (*unstructured.Unstructured, 
 	return c.Resource(SecretResource).Namespace(namespace).Get(ctx, name, metav1.GetOptions{})
 }
 
-func (c *Client) GetSecretBinding(name string) (*SecretBinding, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), requestTimeout)
-	defer cancel()
-	secretBinding, err := c.Resource(SecretBindingResource).Namespace(c.namespace).Get(ctx, name, metav1.GetOptions{})
-	if err != nil {
-		return nil, err
-	}
-	return NewSecretBinding(*secretBinding), err
-}
-
 func (c *Client) GetCredentialsBinding(name string) (*CredentialsBinding, error) {
 	ctx, cancel := context.WithTimeout(context.Background(), requestTimeout)
 	defer cancel()
@@ -65,12 +55,6 @@ func (c *Client) GetCredentialsBinding(name string) (*CredentialsBinding, error)
 		return nil, err
 	}
 	return NewCredentialsBinding(*binding), err
-}
-
-func (c *Client) GetSecretBindings(labelSelector string) (*unstructured.UnstructuredList, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), requestTimeout)
-	defer cancel()
-	return c.Resource(SecretBindingResource).Namespace(c.namespace).List(ctx, metav1.ListOptions{LabelSelector: labelSelector})
 }
 
 func (c *Client) GetCredentialsBindings(labelSelector string) (*unstructured.UnstructuredList, error) {
@@ -83,44 +67,6 @@ func (c *Client) GetShoots() (*unstructured.UnstructuredList, error) {
 	ctx, cancel := context.WithTimeout(context.Background(), requestTimeout)
 	defer cancel()
 	return c.Resource(ShootResource).Namespace(c.namespace).List(ctx, metav1.ListOptions{})
-}
-
-func (c *Client) GetLeastUsedSecretBindingFromSecretBindings(secretBindings []unstructured.Unstructured) (*SecretBinding, error) {
-	usageCount := make(map[string]int, len(secretBindings))
-	for _, s := range secretBindings {
-		usageCount[s.GetName()] = 0
-	}
-
-	shoots, err := c.GetShoots()
-	if err != nil {
-		return nil, fmt.Errorf("while listing shoots: %w", err)
-	}
-
-	if shoots == nil || len(shoots.Items) == 0 {
-		return &SecretBinding{Unstructured: secretBindings[0]}, nil
-	}
-
-	for _, shoot := range shoots.Items {
-		s := Shoot{Unstructured: shoot}
-		count, found := usageCount[s.GetSpecSecretBindingName()]
-		if !found {
-			continue
-		}
-
-		usageCount[s.GetSpecSecretBindingName()] = count + 1
-	}
-
-	min := usageCount[secretBindings[0].GetName()]
-	minIndex := 0
-
-	for i, sb := range secretBindings {
-		if usageCount[sb.GetName()] < min {
-			min = usageCount[sb.GetName()]
-			minIndex = i
-		}
-	}
-
-	return &SecretBinding{Unstructured: secretBindings[minIndex]}, nil
 }
 
 func (c *Client) GetLeastUsedCredentialsBindingFromSecretBindings(credentialsBindings []unstructured.Unstructured) (*CredentialsBinding, error) {
@@ -171,16 +117,6 @@ func (c *Client) UpdateCredentialsBinding(credentialsBinding *CredentialsBinding
 	return NewCredentialsBinding(*u), nil
 }
 
-func (c *Client) UpdateSecretBinding(secretBinding *SecretBinding) (*SecretBinding, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), requestTimeout)
-	defer cancel()
-	u, err := c.Resource(SecretBindingResource).Namespace(c.namespace).Update(ctx, &secretBinding.Unstructured, metav1.UpdateOptions{})
-	if err != nil {
-		return nil, err
-	}
-	return NewSecretBinding(*u), nil
-}
-
 type CredentialsBinding struct {
 	unstructured.Unstructured
 }
@@ -213,40 +149,6 @@ func (b *CredentialsBinding) SetSecretRefName(val string) {
 
 func (b *CredentialsBinding) SetSecretRefNamespace(val string) {
 	_ = unstructured.SetNestedField(b.Unstructured.Object, val, "credentialsRef", "namespace")
-}
-
-type SecretBinding struct {
-	unstructured.Unstructured
-}
-
-func NewSecretBinding(u unstructured.Unstructured) *SecretBinding {
-	return &SecretBinding{u}
-}
-
-func (b *SecretBinding) GetSecretRefName() string {
-	str, _, err := unstructured.NestedString(b.Unstructured.Object, "secretRef", "name")
-	if err != nil {
-		// NOTE this is a safety net, gardener v1beta1 API would need to break the contract for this to panic
-		panic(fmt.Sprintf("SecretBinding missing field '.secretRef.name': %v", err))
-	}
-	return str
-}
-
-func (b *SecretBinding) GetSecretRefNamespace() string {
-	str, _, err := unstructured.NestedString(b.Unstructured.Object, "secretRef", "namespace")
-	if err != nil {
-		// NOTE this is a safety net, gardener v1beta1 API would need to break the contract for this to panic
-		panic(fmt.Sprintf("SecretBinding missing field '.secretRef.namespace': %v", err))
-	}
-	return str
-}
-
-func (b *SecretBinding) SetSecretRefName(val string) {
-	_ = unstructured.SetNestedField(b.Unstructured.Object, val, "secretRef", "name")
-}
-
-func (b *SecretBinding) SetSecretRefNamespace(val string) {
-	_ = unstructured.SetNestedField(b.Unstructured.Object, val, "secretRef", "namespace")
 }
 
 type Shoot struct {
@@ -300,10 +202,8 @@ func (b Shoot) GetSpecRegion() string {
 
 var (
 	SecretResource             = schema.GroupVersionResource{Group: "", Version: "v1", Resource: "secrets"}
-	SecretBindingResource      = schema.GroupVersionResource{Group: "core.gardener.cloud", Version: "v1beta1", Resource: "secretbindings"}
 	ShootResource              = schema.GroupVersionResource{Group: "core.gardener.cloud", Version: "v1beta1", Resource: "shoots"}
 	SecretGVK                  = schema.GroupVersionKind{Group: "", Version: "v1", Kind: "Secret"}
-	SecretBindingGVK           = schema.GroupVersionKind{Group: "core.gardener.cloud", Version: "v1beta1", Kind: "SecretBinding"}
 	ShootGVK                   = schema.GroupVersionKind{Group: "core.gardener.cloud", Version: "v1beta1", Kind: "Shoot"}
 	CredentialsBindingResource = schema.GroupVersionResource{Group: "security.gardener.cloud", Version: "v1alpha1", Resource: "credentialsbindings"}
 	CredentialsBindingGVK      = schema.GroupVersionKind{Group: "security.gardener.cloud", Version: "v1alpha1", Kind: "CredentialsBinding"}
