+++ docs/contributor/03-10-hyperscaler-account-pool.md
@@ -38,7 +38,7 @@ metadata:
 
 For a certain type of SAP BTP, Kyma runtimes, KEB can use the same credentials for multiple tenants.
 In such a case, the Secret with credentials must be labeled differently by adding the **shared** label set to `true`. Shared credentials are not assigned to any tenant.
-Multiple tenants can share the Secret with credentials. That is, many shoots (Shoot resources) can refer to the same Secret. This reference is represented by the SecretBinding (CredentialsBinding) resource.
+Multiple tenants can share the Secret with credentials. That is, many shoots (Shoot resources) can refer to the same Secret. This reference is represented by the CredentialsBinding resource.
 When KEB queries for a Secret for a given hyperscaler, the least used Secret is chosen.  
 
 This is an example of a Kubernetes Secret that stores shared credentials:
