+++ resources/keb/values.yaml
@@ -498,12 +498,6 @@ trialRegionsMapping: |-
 # If true, the broker processes update requests for service instances.
 osbUpdateProcessingEnabled: "true"
 
-# If true, the broker holds any operation with HAP assignments. It is designed for migration (SecretBinding to CredentialBinding).
-holdHAPSteps: "false"
-
-# Name of the Gardener resource, which the broker uses to look up for hyperscaler assignment. Allowed values: SecretBinding or CredentialsBinding.
-subscriptionGardenerResource: SecretBinding
-
 # If true, the broker exposes the API endpoint that returns the availability of machine types.
 machinesAvailabilityEndpoint: false
